decouple webglsampler + attempt at decoupling uniformlocation
