+++ src/lib.rs
@@ -14,8 +14,10 @@ pub mod glenum;
 pub mod query;
 pub mod renderbuffer;
 pub mod rendering_context;
+pub mod sampler;
 pub mod shader_program;
 pub mod texture;
+pub mod uniform_location;
 pub mod vertex_array_object;
 
 pub use buffer::WebGLRSBuffer;
@@ -24,6 +26,8 @@ pub use glenum::*;
 pub use query::WebGLRSQuery;
 pub use renderbuffer::WebGLRSRenderbuffer;
 pub use rendering_context::*;
+pub use sampler::WebGLRSSampler;
 pub use shader_program::{WebGLRSProgram, WebGLRSShader};
 pub use texture::WebGLRSTexture;
+pub use uniform_location::WebGLRSUniformLocation;
 pub use vertex_array_object::WebGLRSVertexArrayObject;
