+++ src/rendering_context.rs
@@ -501,13 +501,6 @@ extern "C" {
     #[wasm_bindgen(method, js_name = enableVertexAttribArray)]
     pub fn enable_vertex_attrib_array(this: &WebGL2RenderingContext, index: u32);
 
-    /// The `WebGLRenderingContext.getUniform()` method of the WebGL API returns the value of a uniform variable
-    /// at a given location.
-    /* FIXME: this method can have a lot of different return types -> figure out what to do
-    #[wasm_bindgen(method, js_name = getUniform)]
-    pub fn get_uniform(this: &WebGL2RenderingContext, program: WebGLProgram, location: WebGLUniformLocation)
-    */
-
     /// The `WebGLRenderingContext.getVertexAttrib()` method of the WebGL API returns information about a vertex
     /// attribute at a given position.
     /* FIXME: a lot of different return value possibilities
@@ -524,168 +517,6 @@ extern "C" {
         index: u32,
         pname: VertexAttrib,
     ) -> i64;
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform1f)]
-    pub fn uniform_1f(this: &WebGL2RenderingContext, location: &WebGLUniformLocation, v0: f32);
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform2f)]
-    pub fn uniform_2f(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        v0: f32,
-        v1: f32,
-    );
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform3f)]
-    pub fn uniform_3f(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        v0: f32,
-        v1: f32,
-        v2: f32,
-    );
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform4f)]
-    pub fn uniform_4f(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        v0: f32,
-        v1: f32,
-        v2: f32,
-        v3: f32,
-    );
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform1i)]
-    pub fn uniform_1i(this: &WebGL2RenderingContext, location: &WebGLUniformLocation, v0: i32);
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform2i)]
-    pub fn uniform_2i(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        v0: i32,
-        v1: i32,
-    );
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform3i)]
-    pub fn uniform_3i(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        v0: i32,
-        v1: i32,
-        v2: i32,
-    );
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform4i)]
-    pub fn uniform_4i(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        v0: i32,
-        v1: i32,
-        v2: i32,
-        v3: i32,
-    );
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform1fv)]
-    pub fn uniform_1fv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        value: Vec<f32>,
-    );
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform2fv)]
-    pub fn uniform_2fv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        value: Vec<f32>,
-    );
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform3fv)]
-    pub fn uniform_3fv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        value: Vec<f32>,
-    );
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform4fv)]
-    pub fn uniform_4fv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        value: Vec<f32>,
-    );
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform1iv)]
-    pub fn uniform_1iv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        value: Vec<i32>,
-    );
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform2iv)]
-    pub fn uniform_2iv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        value: Vec<i32>,
-    );
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform3iv)]
-    pub fn uniform_3iv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        value: Vec<i32>,
-    );
-
-    /// The `WebGLRenderingContext.uniform[1234][fi][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform4iv)]
-    pub fn uniform_4iv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        value: Vec<i32>,
-    );
-
-    /// The `WebGLRenderingContext.uniformMatrix[234]fv()` methods of the WebGL API specify matrix values for
-    /// uniform variables.
-    #[wasm_bindgen(method, js_name = uniformMatrix2fv)]
-    pub fn uniform_matrix_2fv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        transpose: bool,
-        value: Vec<f32>,
-    );
-
-    /// The `WebGLRenderingContext.uniformMatrix[234]fv()` methods of the WebGL API specify matrix values for
-    /// uniform variables.
-    #[wasm_bindgen(method, js_name = uniformMatrix3fv)]
-    pub fn uniform_matrix_3fv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        transpose: bool,
-        value: Vec<f32>,
-    );
-
-    /// The `WebGLRenderingContext.uniformMatrix[234]fv()` methods of the WebGL API specify matrix values for
-    /// uniform variables.
-    #[wasm_bindgen(method, js_name = uniformMatrix4fv)]
-    pub fn uniform_matrix_4fv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        transpose: bool,
-        value: Vec<f32>,
-    );
 
     /// The WebGLRenderingContext.vertexAttrib[1234]f[v]() methods of the WebGL API specify constant values for generic vertex attributes.
     #[wasm_bindgen(method, js_name = vertexAttrib1f)]
@@ -942,146 +773,6 @@ extern "C" {
         height: u32,
     );
 
-    /// The `WebGL2RenderingContext.uniform[1234][uif][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform1ui)]
-    pub fn uniform_1ui(this: &WebGL2RenderingContext, location: &WebGLUniformLocation, v0: u32);
-
-    /// The `WebGL2RenderingContext.uniform[1234][uif][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform2ui)]
-    pub fn uniform_2ui(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        v0: u32,
-        v1: u32,
-    );
-
-    /// The `WebGL2RenderingContext.uniform[1234][uif][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform3ui)]
-    pub fn uniform_3ui(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        v0: u32,
-        v1: u32,
-        v2: u32,
-    );
-
-    /// The `WebGL2RenderingContext.uniform[1234][uif][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform4ui)]
-    pub fn uniform_4ui(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        v0: u32,
-        v1: u32,
-        v2: u32,
-        v3: u32,
-    );
-
-    /// The `WebGL2RenderingContext.uniform[1234][uif][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform1uiv)]
-    pub fn uniform_1uiv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        value: Vec<u32>,
-    );
-
-    /// The `WebGL2RenderingContext.uniform[1234][uif][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform2uiv)]
-    pub fn uniform_2uiv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        value: Vec<u32>,
-    );
-
-    /// The `WebGL2RenderingContext.uniform[1234][uif][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform3uiv)]
-    pub fn uniform_3uiv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        value: Vec<u32>,
-    );
-
-    /// The `WebGL2RenderingContext.uniform[1234][uif][v]()` methods of the WebGL API specify values of uniform variables.
-    #[wasm_bindgen(method, js_name = uniform4uiv)]
-    pub fn uniform_4uiv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        value: Vec<u32>,
-    );
-
-    //TODO all uniform vector methods with optional srcoffset and srclength
-
-    /// The `WebGL2RenderingContext.uniformMatrix[234]x[234]fv()` methods of the WebGL 2 API specify matrix values for
-    /// uniform variables.
-    #[wasm_bindgen(method, js_name = uniformMatrix2x3fv)]
-    pub fn uniform_matrix_2x3fv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        transpose: bool,
-        data: Vec<f32>,
-        srcOffset: u32,
-        srcLength: u32,
-    );
-
-    /// The `WebGL2RenderingContext.uniformMatrix[234]x[234]fv()` methods of the WebGL 2 API specify matrix values for
-    /// uniform variables.
-    #[wasm_bindgen(method, js_name = uniformMatrix2x4fv)]
-    pub fn uniform_matrix_2x4fv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        transpose: bool,
-        data: Vec<f32>,
-        srcOffset: u32,
-        srcLength: u32,
-    );
-
-    /// The `WebGL2RenderingContext.uniformMatrix[234]x[234]fv()` methods of the WebGL 2 API specify matrix values for
-    /// uniform variables.
-    #[wasm_bindgen(method, js_name = uniformMatrix3x2fv)]
-    pub fn uniform_matrix_3x2fv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        transpose: bool,
-        data: Vec<f32>,
-        srcOffset: u32,
-        srcLength: u32,
-    );
-
-    /// The `WebGL2RenderingContext.uniformMatrix[234]x[234]fv()` methods of the WebGL 2 API specify matrix values for
-    /// uniform variables.
-    #[wasm_bindgen(method, js_name = uniformMatrix3x4fv)]
-    pub fn uniform_matrix_3x4fv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        transpose: bool,
-        data: Vec<f32>,
-        srcOffset: u32,
-        srcLength: u32,
-    );
-
-    /// The `WebGL2RenderingContext.uniformMatrix[234]x[234]fv()` methods of the WebGL 2 API specify matrix values for
-    /// uniform variables.
-    #[wasm_bindgen(method, js_name = uniformMatrix4x2fv)]
-    pub fn uniform_matrix_4x2fv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        transpose: bool,
-        data: Vec<f32>,
-        srcOffset: u32,
-        srcLength: u32,
-    );
-
-    /// The `WebGL2RenderingContext.uniformMatrix[234]x[234]fv()` methods of the WebGL 2 API specify matrix values for
-    /// uniform variables.
-    #[wasm_bindgen(method, js_name = uniformMatrix4x3fv)]
-    pub fn uniform_matrix_4x3fv(
-        this: &WebGL2RenderingContext,
-        location: &WebGLUniformLocation,
-        transpose: bool,
-        data: Vec<f32>,
-        srcOffset: u32,
-        srcLength: u32,
-    );
-
     /// The `WebGL2RenderingContext.vertexAttribI4[u]i[v]()` methods of the WebGL 2 API specify integer values for
     /// generic vertex attributes.
     #[wasm_bindgen(method, js_name = vertexAttribI4i)]
@@ -1226,28 +917,6 @@ extern "C" {
         depth: f32,
         stencil: i32,
     );
-
-    /// The `WebGL2RenderingContext.createSampler()` method of the WebGL 2 API creates and initializes
-    /// WebGLSampler objects.
-    #[wasm_bindgen(method, js_name = createSampler)]
-    pub fn create_sampler(this: &WebGL2RenderingContext) -> WebGLSampler;
-
-    /// The `WebGL2RenderingContext.deleteSampler()` method of the WebGL 2 API deletes a given WebGLSampler
-    /// object.
-    #[wasm_bindgen(method, js_name = deleteSampler)]
-    pub fn delete_sampler(this: &WebGL2RenderingContext, sampler: &WebGLSampler);
-
-    /// The `WebGL2RenderingContext.bindSampler()` method of the WebGL 2 API binds a passed WebGLSampler object to
-    /// the texture unit at the passed index.
-    #[wasm_bindgen(method, js_name = bindSampler)]
-    pub fn bind_sampler(this: &WebGL2RenderingContext, unit: u32, sampler: &WebGLSampler);
-
-    /// The WebGL2RenderingContext.isSampler() method of the WebGL 2 API returns true if the passed object is a valid
-    /// WebGLSampler object.
-    #[wasm_bindgen(method, js_name = isSampler)]
-    pub fn is_sampler(this: &WebGL2RenderingContext, sampler: &WebGLSampler);
-
-    //TODO samplerParameter[if] and getSamplerParameter
 
     /// The `WebGL2RenderingContext.fenceSync()` method of the WebGL 2 API creates a new WebGLSync object and inserts
     /// it into the GL command stream.
@@ -1404,20 +1073,6 @@ extern "C" {
     /// The number of bits of precision that can be represented. For integer formats this value is always 0.
     #[wasm_bindgen(method, getter)]
     pub fn precision(this: &WebGLShaderPrecisionFormat) -> u32;
-}
-
-/// WebGLUniformLocation
-#[derive(Clone, Copy)]
-#[wasm_bindgen]
-extern "C" {
-    pub type WebGLUniformLocation;
-}
-
-/// WebGLSampler
-#[derive(Clone, Copy)]
-#[wasm_bindgen]
-extern "C" {
-    pub type WebGLSampler;
 }
 
 /// WebGLSync
